+++ src/commands/decrypt.ts
@@ -1,21 +1,35 @@
 import { GluegunCommand } from 'gluegun';
 
+const description = `Decrypt files from the <target> directory into their original locations.
+If <sources> has been modified, original locations may not match current <sources> configuration.
+Options:   --force (-f)   Skip all prompts.`;
+
 const command: GluegunCommand = {
   name: 'decrypt',
   alias: ['d'],
-  description:
-    'Decrypt files from the <target> directory into their original locations\nIf <sources> has been modified, original locations may not match current <sources> configuration',
-  run: async (toolbox) => {
-    const { config, decryption, filesystem, print, prompt, purge } = toolbox;
+  description,
+  run: async ({
+    config,
+    decryption,
+    filesystem,
+    parameters,
+    print,
+    prompt,
+    purge,
+  }) => {
+    const { options } = parameters;
+    const force = options.force || options.f;
 
     if (!config.target) {
       print.error('No target defined in the configuration file!');
       return;
     }
 
-    const clear = await prompt.confirm(
-      `Discard source files first (${config.sources.join(', ')})?`,
-    );
+    const clear =
+      force ||
+      (await prompt.confirm(
+        `Discard source files first (${config.sources.join(', ')})?`,
+      ));
     if (clear) {
       print.info(`Discarding files (${config.sources.join(', ')}).`);
       purge.discardSources(config.sources);